requirements.txt
lxml>=4.9.0
PyYAML>=6.0
click>=8.0.0
colorama>=0.4.0
jinja2>=3.1.0